requirements.txt
streamlit
python-dotenv>=0.21.0 
fastmcp
langgraph>=0.2.0
langchain>=0.1.0
langchain-openai>=0.1.0
openai>=1.0.0 
pydantic>=2.0.0